Fix: Program using libfuse does not exit properly upon receiving SIGTERM
Issue Details: This issue occurs during reboot when init sends a SIGTERM (signal 15) to a program using libfuse, but the program fails to exit, resulting in an extended reboot time of 3 seconds. Analysis revealed that the functions pthread_setcancelstate and pthread_cancel in libfuse are not implemented. Consequently, when the main thread receives the SIGTERM signal, it does not interrupt the fuse worker threads, which remain blocked on the read system call. Unless the fuse filesystem is accessed or the process is terminated with SIGKILL (signal 9), the worker threads remain alive.

Solution: The decision was made to implement macros for pthread_setcancelstate and pthread_cancel to provide similar functionality:
  - Simulate pthread_setcancelstate by blocking and unblocking SIGUSR1 signals.
  - Simulate pthread_cancel by sending SIGUSR1 signals.

Test: The test results are as follows
  - Successfully built the program using libfuse.
  - Sent SIGTERM to the program using libfuse to confirm that worker threads exit normally from read and the main process terminates as expected.

Change-Id: Ia9c6d6bd83725d981557a3d31495448fc4631178
Signed-off-by: luoqiangwei1 <luoqiangwei1@xiaomi.com>